requirements.txt
numpy
psycopg2-binary
pymongo==3.7.1
redis==3.3.8
msgpack-python==0.5.6
jupyter
jupyter-client
nbconvert
sqlalchemy
pandas
scipy
scikit-learn
matplotlib
plotly
pillow
nltk==3.4.4
implicit
requests